GfxFontDict: merge reference generation from xpdf 4.00
The GfxFontDict constructor generates a fake indirect reference if the
font dictionary doesn't have a real indirect reference.  It sometimes
assigns the same reference to two different fonts leading to a wrong
font being used.  XPDF 4.00 fixes this by using the hash of the font
data to create the fake reference.

https://bugs.freedesktop.org/show_bug.cgi?id=104565