Fix PDFDoc::checkHeader() for PDFs smaller than 1 KiB
The fix for bug 104502 made it so PDFDoc::checkHeader() would print a
warning and return immediatelly if it encounters an EOF while reading
the first 1024 bytes.

Some PDF files can be smaller than 1024 bytes, for example those used
by pdf2djvu's test suite. The latter would fail due to the unexpected
warnings.

Change the behavior of PDFDoc::checkHeader() when encountering an EOF
so it processes the data it has read so far instead of aborting early.

https://bugs.freedesktop.org/show_bug.cgi?id=105674