Annot: Move appearBuf construction to a new class
Previously we had appearBuf construction functions in Annot itself which was a bit
confusing. Some functions like writeString took appearBuf as parameter, some
other did not. By moving it out to a new class and making all the input parameters const
pointers we make it clear the construction of appearBuf doesn't have any other side
effect on Annot itself or any other class

I'm leaving the functions in their old place to minimize the diff so it's easier to read.
I'll move them in a follow up commit