Replace the implementation of GooString by std::string but keep the exact interface intact.
The approach is slightly different to GooList as it reimplements GooString in terms of
std::string but keeps its interface intact and does expose any std::string functionality
as of now.

This is done since GooString has a significantly larger API surface and exposing both API
would be quite confusing with with some overloads from GooString and some from std::string
being visible. But it does mean we can align the API (they are already pretty close) and
expose new things like a reserve method piece by piece.

It also already helps in that the implementation of GooString is gone except for the
original parts, i.e. the formatting, and we have zero cost conversion from/to std::string
which should help in making more use of it elsewhere. It also gives us do-nothing-access
to optimizations done for the standard library, e.g. word-level implementations of
cmp and friends.

(Note that the resulting GooString.cc is a bit funny as the formatting helper are now
local to the translation unit, but I had to redeclare them to keep the diff small as
I did not change them at all. But if this is done, they could probably just be moved to
where the declarations are to make the source file more readable.)