pdf-inspector: Link to the right libraries
pdf-inspector doesn't use libpoppler-glib, but linked to it.
Changing it to link only to libpoppler revealed that it
used lots of symbols from poppler-glib that poppler-glib
should not export. Fix that by adding the necessary sources
to the pdf-inspector sources and linking to the libraries
needed by these sources (freetype, pthreads).