support 'de facto' tooltip feature
Most pdf readers implement a tooltip feature by
showing the string content of 'TU' field of a
widget annotation that is not linked to any
form field.

Normally, widget annotations carry a reference to a
form field which are used together to implement the
different form widgets. But, the PDF spec does not
forbid standalone (i.e. not linked to any form field)
widget annotations, and the fact is they're been used
by most pdf readers to show a tooltip when the area
of that AnnotWidget is hovered.

Some API added for this feature:

bool FormField::isStandAlone()
void FormField::setStandAlone (bool value)

A standalone FormField means it's not part of Catalog's
Field array, because of that we store them in a new
member inside Page class:

std::vector<FormField*> standaloneFields;

and send them alongside the rest of FormWidgets in the
existant API:

FormPageWidgets *Page::getFormWidgets();

Poppler issue #34

Evince issue:
https://gitlab.gnome.org/GNOME/evince/issues/842