cleanup displayprofile initialization
There were a bunch of global variables that were used to initilize the first version
of the display profiles. This code was removed, and all the static initilization was moved
from GfxColorSpace to GfxState. Furthermore, for most "users" the setting of the
display profile was moved from the static GfxColorSpace::setDisplayProfile function
to the OutputDev class. The latter is now invoked early in the initilization of Gfx
to set the initial state in the GfxState instance.