poppler-glib: fix adding annots in rotated pages
This commit adds support to poppler-glib to
correctly add annotations to pages that are
rotated.

After this commit, annotations added through
Evince will be correctly positioned on rotated
pages (without any code change needed in Evince),
both for normal and 'flagNoRotate' type
annots and also for rotated pages that have a
cropbox set in.

As per PDF spec, annotations in rotated pages
must be saved un-rotated, and the pdf client
will display them rotated according to the
rotation of the page they're in.

Poppler-glib was not un-rotating them when
saving them in core poppler, we now do it
after this change, and poppler-glib will
continue to serve them to apps with page's
rotation applied, through the
poppler_page_get_annot_mapping() API.

No new API has been added or changed as of
this commit.

Evince issue:
https://gitlab.gnome.org/GNOME/evince/-/issues/1385

Poppler issue #256