JBIG2Reader: Fix binary constant
When comparing the file JBIG2Reader.cc with what is in xpdf-4.02,
I noticed that a particular constant used in the method
readGenericBitmap differs in poppler and xpdf.  I have no idea
what that constant is good for, but in a comment both poppler
and xpdf give the binary representation of that constant.
That binary number is identical in poppler and xpdf, and it
matches the value in xpdf, but is off-by-one in poppler.
This leads me to believe that the value in poppler wrong,
possibly a typo.