TextOutputDev: require more spacing between columns
Require more spacing for adjacent text to be
considered a separate column of text.

We do that by increasing 'minColSpacing1' parameter,
which marks the distance, within which, an adjacent
word will be pulled to the current block.

We provide a way to tweak the default value:
double getMinColSpacing1();
void setMinColSpacing1(double val);

Fixes issue #1093