pdftohtml: Reduce sensitivity of duplicate detection
fixes #1117

In some fonts, strings such as "ll" or "ff" are placed close enough
together to trigger duplicate detection in pdftohtml. This commit makes
the detection algorithm less sensitive to reduce the false positives
while still maintaining the original function of the code.

Prior to this commit, if a character's `xMax` is less than 20% of its
height away from the following character's `xMax`, it is treated as a
duplicate and removed. This commit changes that value to 17.5%, which
will reduce the number of false positives without introducing too many
false negatives.