Improve support for custom stamp annotations
This commit improves Poppler support for custom stamp
annotations, by adding a new class called AnnotStampImageHelper in
Poppler core.

The new class takes image data and create an Image
XObject in the document, the AnnotStamp class has been modified to
support the new helper class and to reference the created XObject.

This new implementation has been exposed in the qt5 wrapper as well as the
qt6 one, in which the extraction of the QImage data has been handled.

A new API for preserving the annotation AP stream has been exposed using
the qt wrapper as well, so that users are able to temporarily store it.