Update pdftohtml duplicate detection
The delta values used now are the same as the ones used in
pdftotext, which have proven to be much more reliable.
Additionally the search range on the xaxis for duplicate strings has
been increased, which seems to vastly improve the ability to find
duplicates. This algorithm can now properly detect duplicates as shown
in #321.