glib: Add poppler_document_new_from_fd
While it's already possible to create a PopplerDocument
for STDIN by using fd://0 as the URI, it was not yet possible
to create a PopplerDocument from a file descriptor.

This adds poppler_document_new_from_fd(), which accepts
a readable FD for a regular file, or for STDIN.

Add a --fd option to test/gtk-test to test this. When used,
gtk-test arguments are FD numbers instead of filenames or URIs.
To test, use e.g.

$ 3<test.pdf ./gtk-test --fd 3