Use new/delete for class type3_font_info_t
In a later commit I want to change type3_font_info_t and make it
contain types with non-trivial destructors.  However, these
destructors are only called when the destructor of type3_font_info_t
is called.  And this only happens if type3_font_info_t is deleted
with 'delete', not with 'free'.