Remove need for dual signing of data
Instead of creating a placeholder signature of dummy data, just use
dummy data as a placeholder and then pad the real signature with zeroes
to fit in the same size as the dummy data.

This is also kind of what the pdf specification suggests.

It also ensures that for certain smart card setups requiring a distinct
pin per signature will only be asked once.