Fix crash when signing existing signature
On windows, helvetica is not available and that makes poppler crash, so
fix crash and add fallback to arial.

This has already been implemented for "new" signatures, both the
fallback and the crash in
09fdfecea3a13d30b5c52e1258d17549739d9ea8 and
a563801e1a6be5e70645472d82f4ba8534b454d1

Refactor the fallback code to be reusable and use the same crash guard.