Version symbols in poppler core
This prevents crashes and collisions in the case where two poppler core
libraries are in the same process.

The likely case is if an application is linked to both poppler core
and a stable frontend, and poppler is updated and the application
is not yet recompiled against the newer poppler core, then the newer
frontend pulls in newer poppler core while the application pulls in the
older poppler core and that leads to crashes. In general, this can be
fixed by versioning the symbols to prevent mix and match of symbols.

Patch by Andreas Metzler <ametzler@bebt.de>