Move the method sanitizedName from GooString to PDFDoc
Because it is only used by the PDFDoc class, and it does not seem
to be generic enough for a GooString method.

Also, this patch modifies the method to return the string as a
std::string object by value, rather than as a pointer to a
heap-allocated GooString object.  This saves one heap allocation.