Object: Add constructor from r-value std::string
This can simplify the calling code when the string to use
is stack-allocated.

Unfortunately, the new constructor has to do a heap allocation
internally, because Object stores a pointer to a string,
not a string value.

With the new constructor, construction of Object objects from
std::initializer_list is now ambiguous. This leads to compiler
errors in a few places where {num, gen} is used to initialize
an Object with a Ref object.  The patch replaces this construction
by Ref{num, gen}, which fixes the problem.