Enable pdfimages to print filenames to stdout.
Add a command line option -print-filenames so that after writing each
imaage to file, pdfimages will print the filename to stdout. This is in
line with the --batch-print option of scanimage and the stream_filelist
parameter of tesseract. It allows easy parallelization of different
stages of an image processing pipeline.