Move PDFRectangle to a separate header file, make methods constexpr
The class is very lightweight, without any further dependencies. Making
it available outside of Page.h allows to consolidate several cases where
e.g. bounding boxes are declared as a quadruple of doubles, while
PDFRectangle would be a much better fit - most notably also for several
methods which use four pointers to double to return a BBox.

All methods of PDFRectangle can be constexpr, including the constructors.
This also allows it to test it at compile time using static_assert.

All methods but clipTo were already implicitly inline. It is
slightly more complex than the other methods, but having it constexpr
is a requirement for testing it via static_assert.